Move layout code into sub-package
