+++ openapi_spec_tools/cli_gen/cli.py
@@ -31,18 +31,18 @@ from openapi_spec_tools.cli_gen.files import find_unreferenced
 from openapi_spec_tools.cli_gen.files import generate_node
 from openapi_spec_tools.cli_gen.files import generate_tree_file
 from openapi_spec_tools.cli_gen.files import generate_tree_node
-from openapi_spec_tools.cli_gen.layout import DEFAULT_START
-from openapi_spec_tools.cli_gen.layout import check_pagination_definitions
-from openapi_spec_tools.cli_gen.layout import file_to_tree
-from openapi_spec_tools.cli_gen.layout import open_layout
-from openapi_spec_tools.cli_gen.layout import operation_duplicates
-from openapi_spec_tools.cli_gen.layout import operation_order
-from openapi_spec_tools.cli_gen.layout import subcommand_missing_properties
-from openapi_spec_tools.cli_gen.layout import subcommand_order
-from openapi_spec_tools.cli_gen.layout import subcommand_references
-from openapi_spec_tools.cli_gen.layout_generator import LayoutGenerator
-from openapi_spec_tools.cli_gen.layout_generator import write_layout
-from openapi_spec_tools.cli_gen.layout_types import LayoutNode
+from openapi_spec_tools.layout.layout_generator import LayoutGenerator
+from openapi_spec_tools.layout.layout_generator import write_layout
+from openapi_spec_tools.layout.types import LayoutNode
+from openapi_spec_tools.layout.utils import DEFAULT_START
+from openapi_spec_tools.layout.utils import check_pagination_definitions
+from openapi_spec_tools.layout.utils import file_to_tree
+from openapi_spec_tools.layout.utils import open_layout
+from openapi_spec_tools.layout.utils import operation_duplicates
+from openapi_spec_tools.layout.utils import operation_order
+from openapi_spec_tools.layout.utils import subcommand_missing_properties
+from openapi_spec_tools.layout.utils import subcommand_order
+from openapi_spec_tools.layout.utils import subcommand_references
 from openapi_spec_tools.types import OasField
 from openapi_spec_tools.utils import remove_property
 from openapi_spec_tools.utils import remove_schema_tags
