+++ openapi_spec_tools/cli_gen/files.py
@@ -11,7 +11,7 @@ from openapi_spec_tools.cli_gen._tree import TreeField
 from openapi_spec_tools.cli_gen._tree import TreeNode
 from openapi_spec_tools.cli_gen.cli_generator import CliGenerator
 from openapi_spec_tools.cli_gen.constants import GENERATOR_LOG_CLASS
-from openapi_spec_tools.cli_gen.layout_types import LayoutNode
+from openapi_spec_tools.layout.types import LayoutNode
 from openapi_spec_tools.types import OasField
 from openapi_spec_tools.utils import map_operations
